+++ src/main/java/com/emprestimo/emprestimoapp/controller/CustomerController.java
@@ -35,13 +35,13 @@ public class CustomerController {
 	}
 
 	@GetMapping("/edit/{id}")
-	public ModelAndView edit(@PathVariable("id") Integer id) {
+	public ModelAndView edit(@PathVariable("id") Long id) {
 
 		return add(customerService.findOne(id));
 	}
 
 	@GetMapping("/delete/{id}")
-	public ModelAndView delete(@PathVariable("id") Integer id) {
+	public ModelAndView delete(@PathVariable("id") Long id) {
 
 		customerService.delete(id);
